refactor: move to futures.rs
